+++ lib/src/sliver/sliver_flexible_table_content.dart
@@ -1,4 +1,5 @@
 import 'package:flexible_scrollable_table_view/src/arguments/table_build_arguments.dart';
+import 'package:flexible_scrollable_table_view/src/layout_builder/lazy_sliver_layout_builder.dart';
 import 'package:flexible_scrollable_table_view/src/widgets/flexible_table_content.dart';
 import 'package:flutter/widgets.dart';
 
@@ -15,30 +16,34 @@ class SliverFlexibleTableContent<T> extends FlexibleTableContent<T> {
 
   @override
   Widget build(BuildContext context) {
-    final AbsTableBuildArguments<T> arguments = TableBuildArguments<T>(
-      controller,
-      configurations,
-    );
-    return ValueListenableBuilder<List<T>>(
-      valueListenable: controller,
-      builder: (context, value, child) {
-        final double? itemExtent = value.isEmpty ? null : super.itemExtent;
-        final int totalItemCount = getItemCount(value);
-        final SliverChildDelegate delegate = SliverChildBuilderDelegate(
-          (context, index) => buildItem(
-            context,
-            arguments: arguments,
-            value: value,
-            index: index,
-            itemCount: totalItemCount,
-            rowWidth: null,
-          ),
-          childCount: totalItemCount,
+    return LazySliverLayoutBuilder(
+      builder: (context, parentWidth) {
+        final AbsTableBuildArguments<T> arguments = TableBuildArguments<T>(
+          controller,
+          configurations,
+          parentWidth,
         );
-        if (itemExtent == null) {
-          return SliverList(delegate: delegate);
-        }
-        return SliverFixedExtentList(delegate: delegate, itemExtent: itemExtent);
+        return ValueListenableBuilder<List<T>>(
+          valueListenable: controller,
+          builder: (context, value, child) {
+            final double? itemExtent = value.isEmpty ? null : super.itemExtent;
+            final int totalItemCount = getItemCount(value);
+            final SliverChildDelegate delegate = SliverChildBuilderDelegate(
+              (context, index) => buildItem(
+                context,
+                arguments: arguments,
+                value: value,
+                index: index,
+                itemCount: totalItemCount,
+              ),
+              childCount: totalItemCount,
+            );
+            if (itemExtent == null) {
+              return SliverList(delegate: delegate);
+            }
+            return SliverFixedExtentList(delegate: delegate, itemExtent: itemExtent);
+          },
+        );
       },
     );
   }