+++ lib/src/widgets/flexible_table_content.dart
@@ -2,10 +2,10 @@ import 'package:flexible_scrollable_table_view/src/addition/flexible_table_addit
 import 'package:flexible_scrollable_table_view/src/animation/flexible_table_animations.dart';
 import 'package:flexible_scrollable_table_view/src/arguments/table_build_arguments.dart';
 import 'package:flexible_scrollable_table_view/src/arguments/table_row_build_arguments.dart';
-import 'package:flexible_scrollable_table_view/src/constraint/lazy_row_layout_builder.dart';
 import 'package:flexible_scrollable_table_view/src/decoration/flexible_table_decorations.dart';
 import 'package:flexible_scrollable_table_view/src/flexible_table_configurations.dart';
 import 'package:flexible_scrollable_table_view/src/flexible_table_controller.dart';
+import 'package:flexible_scrollable_table_view/src/layout_builder/lazy_layout_builder.dart';
 import 'package:flexible_scrollable_table_view/src/sliver/sliver_flexible_table_content.dart';
 import 'package:flutter/widgets.dart';
 
@@ -25,7 +25,6 @@ class FlexibleTableContent<T> extends StatelessWidget {
     this.primary,
     this.verticalPhysics,
     this.horizontalPhysics,
-    this.verticalScrollable = true,
   });
 
   factory FlexibleTableContent.sliver(
@@ -58,14 +57,13 @@ class FlexibleTableContent<T> extends StatelessWidget {
   final ScrollPhysics? verticalPhysics;
   final ScrollPhysics? horizontalPhysics;
 
-  final bool verticalScrollable;
-
   bool get hasHeader => additions?.headerBuilder != null;
 
   bool get hasFooter => additions?.footerBuilder != null;
 
   bool get hasPlaceholder => additions?.placeholderBuilder != null;
 
+  @protected
   int getItemCount(List<T> dataList) {
     //当数据列是空的
     if (dataList.isEmpty) {
@@ -76,28 +74,31 @@ class FlexibleTableContent<T> extends StatelessWidget {
     return (hasHeader ? 1 : 0) + dataList.length + (hasFooter ? 1 : 0);
   }
 
+  @protected
   bool isHeaderIndex(int index) {
     //当下标为 0，并且又有头部的时候，则当前下标为头部
     return index == 0 && hasHeader;
   }
 
+  @protected
   bool isFooterIndex(List<T> value, int index) {
     //当下标 = 数据长度 + 头部数量，则当前下标为尾部
     return index == value.length + (hasHeader ? 1 : 0);
   }
 
+  @protected
   int realDataIndex(int index) {
     //真实的数据下标 = 列表下标 - 头部数量
     return index - (hasHeader ? 1 : 0);
   }
 
+  @protected
   Widget buildItem(
     BuildContext context, {
     required AbsTableBuildArguments<T> arguments,
     required List<T> value,
     required int index,
     required int itemCount,
-    required double? rowWidth,
   }) {
     //数据是空的，却又需要构建列表项，说明是需要绘制占位组件
     if (value.isEmpty && hasPlaceholder) {
@@ -109,32 +110,9 @@ class FlexibleTableContent<T> extends StatelessWidget {
     if (isFooterIndex(value, index)) {
       return additions!.footerBuilder!.call(arguments);
     }
-    final int dataIndex = realDataIndex(index);
-    if (rowWidth == null) {
-      return LazyRowLayoutBuilder(
-        rowHeight: configurations.rowHeight.getInfoRowHeight(
-          controller,
-          dataIndex,
-          value[dataIndex],
-        ),
-        builder: (context, parentWidth) => FlexibleTableInfoRow<T>(
-          arguments.toInfoRowArguments(
-            rowWidth: parentWidth,
-            dataIndex: dataIndex,
-            dataList: value,
-            currentItemIndex: index,
-            totalItemCount: itemCount,
-          ),
-          decorations: decorations,
-          animations: animations,
-          physics: horizontalPhysics,
-        ),
-      );
-    }
     return FlexibleTableInfoRow<T>(
       arguments.toInfoRowArguments(
-        rowWidth: rowWidth,
-        dataIndex: dataIndex,
+        dataIndex: realDataIndex(index),
         dataList: value,
         currentItemIndex: index,
         totalItemCount: itemCount,
@@ -145,6 +123,7 @@ class FlexibleTableContent<T> extends StatelessWidget {
     );
   }
 
+  @protected
   double? get itemExtent {
     if (configurations.rowHeight.infoRowHeightBuilder != null || configurations.rowHeight.fixedInfoRowHeight == null) {
       return null;
@@ -164,13 +143,14 @@ class FlexibleTableContent<T> extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final AbsTableBuildArguments<T> arguments = TableBuildArguments<T>(
-      controller,
-      configurations,
-    );
-    return LayoutBuilder(
+    return LazyLayoutBuilder(
       builder: (context, constraints) {
-        Widget child = ValueListenableBuilder<List<T>>(
+        final AbsTableBuildArguments<T> arguments = TableBuildArguments<T>(
+          controller,
+          configurations,
+          constraints.maxWidth,
+        );
+        final Widget child = ValueListenableBuilder<List<T>>(
           valueListenable: controller,
           builder: (context, value, child) {
             final int totalItemCount = getItemCount(value);
@@ -182,19 +162,18 @@ class FlexibleTableContent<T> extends StatelessWidget {
               padding: EdgeInsets.zero,
               scrollDirection: Axis.vertical,
               itemExtent: value.isEmpty ? null : itemExtent,
-              physics: verticalScrollable ? verticalPhysics : const NeverScrollableScrollPhysics(),
+              physics: verticalPhysics,
               itemBuilder: (context, index) => buildItem(
                 context,
                 arguments: arguments,
                 value: value,
                 index: index,
                 itemCount: totalItemCount,
-                rowWidth: constraints.maxWidth,
               ),
             );
           },
         );
-        if (!verticalScrollable) {
+        if (shrinkWrap) {
           return SizedBox(
             width: constraints.maxWidth,
             child: child,